moved types to environment.ts
