+++ packages/core/src/index.ts
@@ -58,6 +58,7 @@ import { PacketForwardInteraction } from './interactions/packet/forward'
 
 import { Packet } from './messages'
 import { localAddressesFirst, AddressSorter, backoff, durations, isErrorOutOfFunds } from '@hoprnet/hopr-utils'
+import type { ResolvedEnvironment } from './environment'
 
 const log = Debug(`hopr-core`)
 const verbose = Debug('hopr-core:verbose')
@@ -70,39 +71,6 @@ interface NetOptions {
 type PeerStoreAddress = {
   id: PeerId
   multiaddrs: Multiaddr[]
-}
-
-export type Network = {
-  id: string
-  description: string
-  chain_id: number // >= 0
-  live: boolean
-  default_provider: string // a valid HTTP url pointing at a RPC endpoint
-  gas: string // e.g. '1 gwei'
-  gas_multiplier: number // e.g. 1.1
-  native_token_name: string
-  hopr_token_name: string
-}
-
-export type Environment = {
-  id: string
-  network_id: string // must match one of the Network.id
-  channel_contract_deploy_block: number // >= 0
-  token_contract_address: string // an Ethereum address
-  channels_contract_address: string // an Ethereum address
-}
-
-export type ProtocolConfig = {
-  environments: Environment[]
-  networks: Network[]
-}
-
-export type ResolvedEnvironment = {
-  id: string
-  network: Network
-  channel_contract_deploy_block: number
-  token_contract_address: string // an Ethereum address
-  channels_contract_address: string // an Ethereum address
 }
 
 export type HoprOptions = {
@@ -997,3 +965,4 @@ export { Hopr as default, LibP2P }
 export * from './constants'
 export { PassiveStrategy, PromiscuousStrategy, SaneDefaults, findPath }
 export type { ChannelsToOpen, ChannelsToClose }
+export type { ProtocolConfig } from './environment'
